Renamed the new interp methods to a more consistent pattern.
